fix - deleted client manager as there was no point in its existence. Now there are receivers instead of contend handlers. Those classes and interface are soon to be renamed
