+++ client/src/main/java/clientModules/response/receivers/CommandExecutionResultReceiver.java
@@ -1,0 +1,11 @@
+package clientModules.response.receivers;
+
+import responses.Response;
+
+public class CommandExecutionResultReceiver implements ResponseContentHandler {
+
+    @Override
+    public void handleResponseContent(Response response) {
+
+    }
+}
