--- client/src/main/java/clientModules/response/responseContentHandlers/ResponseOutputHandler.java
@@ -1,11 +1,0 @@
-package clientModules.response.responseContentHandlers;
-
-import responses.Response;
-
-public class ResponseOutputHandler implements ResponseContentHandler {
-
-    @Override
-    public void handleResponseContent(Response response) {
-
-    }
-}
