+++ server/src/main/java/serverModules/connection/ConnectionModule.java
@@ -10,7 +10,7 @@ import java.net.InetAddress;
 import java.net.SocketException;
 
 public class ConnectionModule implements ReceiveDataAble<RequestData>, SendDataAble {
-    private final int BYTE_SIZE = (int) Math.pow(2, 10);
+    private final int BYTE_SIZE = 1024;
     private final DatagramSocket socket;
 
     protected ConnectionModule(int port) throws SocketException {
