+++ server/src/main/java/validators/LocationValidator.java
@@ -18,10 +18,6 @@ public class LocationValidator implements ValidateAble<Location> {
 
     @Override
     public boolean validate(Location location) {
-        if (location != null) {
-            return location.getX() != null && location.getY() != null
-                    && !Objects.equals(location.getName(), "");
-        }
-        return true;
+        return location == null ? true : (location.getX() != null && location.getY() != null && !Objects.equals(location.getName(), ""));
     }
 }
